% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aedes.R
\name{aedes}
\alias{aedes}
\alias{import_aedes_csv}
\alias{export_aedes_csv}
\alias{aedes_expected}
\title{Tools for working with Aedes challenge forecasts}
\usage{
import_aedes_csv(file, add_vars = NULL)

export_aedes_csv(x, filename = NULL, overwrite = F)

aedes_expected()
}
\arguments{
\item{file}{A csv file formatted for the EPI Aedes Forecasting Challenge.}
}
\value{
A \code{predx} data frame.
}
\description{
These functions are used to convert an Epidemic Prediction Initiative Aedes Forecasting Challenge-formatted csv file to a \code{predx} data frame.
}
\examples{
csv_tempfile <- tempfile()
write.csv(aedesdemo, csv_tempfile, row.names=F)
import_aedes_csv(csv_tempfile)
}
