% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aedesdemo-data.R
\docType{data}
\name{aedesdemo}
\alias{aedesdemo}
\title{Example Aedes forecast}
\format{
A data frame input from the original submission CSV with additional columns for due_date, month, and team.
}
\usage{
data(aedesdemo)
}
\description{
Null forecast the 2019 Aedes Forecasting Challenge
}
\keyword{dataset}
