% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinCat.R
\docType{class}
\name{BinCat-class}
\alias{BinCat-class}
\title{An S4 class to represent a categorical probability distribution.}
\description{
An S4 class to represent a categorical probability distribution.
}
\section{Slots}{

\describe{
\item{\code{predx}}{data.frame with two columns: cat (character) and prob (numeric).}
}}

