% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_errors.R
\name{check_conversion_errors}
\alias{check_conversion_errors}
\alias{collect_error}
\title{Title}
\usage{
check_conversion_errors(x, msg = F)

collect_error(x)
}
\arguments{
\item{x}{list of predx objects.}
}
\description{
Title
}
\section{Functions}{
\itemize{
\item \code{collect_error}: function to collect individual errors
}}

