% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gte0.R
\name{check_gte0}
\alias{check_gte0}
\title{Check that all probabilities are positive}
\usage{
check_gte0(x, colname)
}
\arguments{
\item{probs}{vector of probabilities for a single forecast}
}
\value{
TRUE or error message
}
\description{
Check that all probabilities are positive
}
