% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lte1.R
\name{check_lte1}
\alias{check_lte1}
\title{Check that all probabilities are positive}
\usage{
check_lte1(x, colname)
}
\arguments{
\item{probs}{vector of probabilities for a single forecast}
}
\value{
TRUE or error message
}
\description{
Check that all probabilities are positive
}
