% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_probs_bins_length_match.R
\name{check_probs_bins_length_match}
\alias{check_probs_bins_length_match}
\title{Check that number of forecast bins matches number of probabilties.}
\usage{
check_probs_bins_length_match(probs, bins)
}
\arguments{
\item{bins}{}
}
\value{
TRUE or error message
}
\description{
Check that number of forecast bins matches number of probabilties.
}
