% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_probs_lt1.R
\name{check_probs_lt1}
\alias{check_probs_lt1}
\title{Check that all probabilities are positive}
\usage{
check_probs_lt1(probs)
}
\arguments{
\item{probs}{vector of probabilities for a single forecast}
}
\value{
TRUE or error message
}
\description{
Check that all probabilities are positive
}
