% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_value.R
\name{check_single_value}
\alias{check_single_value}
\title{Check that all probabilities are positive}
\usage{
check_single_value(x)
}
\arguments{
\item{probs}{vector of probabilities for a single forecast}
}
\value{
TRUE or error message
}
\description{
Check that all probabilities are positive
}
