% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_uniform_bin_size.R
\name{check_uniform_bin_size}
\alias{check_uniform_bin_size}
\title{Check that all bins have the same width.}
\usage{
check_uniform_bin_size(bins)
}
\arguments{
\item{bins}{}
}
\value{
TRUE or error message
}
\description{
Check that all bins have the same width.
}
