% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_json}
\alias{export_json}
\title{Export a \code{predx} data frame as a \code{predx}-formatted JSON file}
\usage{
export_json(x, filename = NA, overwrite = F)
}
\arguments{
\item{filename}{character. Filename to write JSON file. If NA, JSON is returned.}

\item{overwrite}{boolean. Overwrite the file if it already exists.}

\item{forecast}{A \code{predx} data frame.}
}
\value{
Silently saves a \code{predx}-formatted JSON if \code{filename} is specified. Otherwise returns a string in the format of the \code{predx} JSON.
}
\description{
Export a \code{predx} data frame as a \code{predx}-formatted JSON file
}
\examples{
predx_demo <- as.predx_df(list(
 location = c('Mercury', 'Venus', 'Earth'),
 target = 'habitability',
 predx = list(Binary(1e-4), Binary(1e-4), Binary(1))
))
json_tempfile <- tempfile()
export_json(predx_demo, json_tempfile)
import_json(json_tempfile)
}
