% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flusight.R
\name{flusight}
\alias{flusight}
\alias{import_flusight_csv}
\alias{export_flusight_csv}
\alias{prep_flusight}
\alias{flusight_ilinet_expected}
\alias{flusight_state_ilinet_expected}
\alias{flusight_hospitalization_expected}
\alias{recode_flusight_bin_end_notincl}
\title{Tools for working with FluSight forecasts}
\usage{
import_flusight_csv(file)

export_flusight_csv(x, dir = NULL, overwrite = F)

prep_flusight(x)

flusight_ilinet_expected()

flusight_state_ilinet_expected()

flusight_hospitalization_expected()

recode_flusight_bin_end_notincl(x)
}
\arguments{
\item{file}{a csv file formatted for the FluSight forecasting challenge}
}
\description{
These functions are used to convert original FluSight-formatted csv and JSON files to `predx_df` objects.
}
\examples{
x = import_flusight_csv('vignettes/EW42-Hist-Avg-2018-10-29.csv')
}
