% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flusightdemo-data.R
\docType{data}
\name{flusightdemo}
\alias{flusightdemo}
\title{Example FluSight forecast}
\format{A data frame input from the original submission CSV.}
\source{
https://github.com/cdcepi/FluSight-forecasts/tree/master/2018-2019/Hist-Avg
}
\usage{
data(flusightdemo)
}
\description{
National-level forecasts for the "Hist-Avg" forecast from Epidemiological Week 42 in 2018
}
\keyword{dataset}
