% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cats.R
\name{get_cats}
\alias{get_cats}
\title{Get vector of categories for a BinCat or SampleCat class predx object}
\usage{
get_cats(x)
}
\arguments{
\item{x}{an object of class BinCat or SampleCat OR a list of those objects}
}
\value{
character vector of categories
}
\description{
Get vector of categories for a BinCat or SampleCat class predx object
}
