% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predx.R
\name{get_predx}
\alias{get_predx}
\title{Generic function for extracting predx objects as base R data structures}
\usage{
get_predx(x, ...)
}
\arguments{
\item{object}{}
}
\description{
Generic function for extracting predx objects as base R data structures
}
