% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predx_colnames.R
\name{get_predx_colnames}
\alias{get_predx_colnames}
\title{Internal function returns the names of columns used for conversion to predx objects}
\usage{
get_predx_colnames(classes)
}
\arguments{
\item{classes}{A vector of predx classes names (e.g. 'Point', 'Binary').}
}
\value{
`predx` tibble
}
\description{
Internal function returns the names of columns used for conversion to predx objects
}
\examples{
get_predx_colnames(c("Binary", "Point", "BinCat"))
}
