% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{Import predx formatted csv}
\usage{
import_csv(file = NULL)
}
\arguments{
\item{file}{A csv file/path *or* a data.frame object. Either should have a `predx_class` column and whatever columns are needed for the incldued classes (e.g. point, bin, prob).}
}
\value{
`predx` tibble
}
\description{
Import predx formatted csv
}
\examples{
import_csv('vignettes/fcast_demo.csv')
}
