% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.R, R/PointCat.R
\docType{class}
\name{Point-class}
\alias{Point-class}
\alias{Point}
\alias{is.Point}
\alias{predx_to_json,Point-method}
\alias{as.data.frame,Point-method}
\alias{transform_predx,Point-method}
\alias{predx_to_json,PointCat-method}
\title{Point class: Point predictions}
\usage{
Point(x)

is.Point(object)

\S4method{predx_to_json}{Point}(x)

\S4method{as.data.frame}{Point}(x, row.names = NULL, optional = FALSE, ...)

\S4method{transform_predx}{Point}(x, to_class, ...)

\S4method{predx_to_json}{PointCat}(x)
}
\description{
A \code{predx} class for point predictions.
}
\details{
\code{Point} objects contain a single numeric point prediction with no other restrictions.

In JSON and CSV representations, this value is named \code{point}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{A single numeric point prediction.}
}}

\examples{

}
