% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointCat.R
\docType{class}
\name{PointCat-class}
\alias{PointCat-class}
\alias{PointCat}
\alias{is.PointCat}
\alias{as.list,PointCat-method}
\alias{as.data.frame,PointCat-method}
\alias{transform_predx,PointCat-method}
\title{PointCat class: Point predictions for categorical variables}
\usage{
PointCat(x)

is.PointCat(object)

\S4method{as.list}{PointCat}(x, ...)

\S4method{as.data.frame}{PointCat}(x, row.names = NULL,
  optional = FALSE, ...)

\S4method{transform_predx}{PointCat}(x, to_class, ...)
}
\description{
A \code{predx} class for point predictions for categorical variables.
}
\details{
\code{PointCat} objects contain a single point prediction of type character
with no other restrictions.

In JSON and CSV representations, this value is named \code{point}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{A single point prediction of type character.}
}}

\examples{

}
