% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_to_json.R
\docType{methods}
\name{predx_to_json}
\alias{predx_to_json}
\alias{predx_to_json,list-method}
\title{Internal code to convert \code{predx} to JSON friendly representation.}
\usage{
predx_to_json(x)

\S4method{predx_to_json}{list}(x)
}
\arguments{
\item{x}{a \code{predx} object or list of \code{predx} objects}
}
\value{
A \code{predx} object or list of \code{predx} objects, with NAs and warning messages for failed conversions.
}
\description{
Internal code to convert \code{predx} to JSON friendly representation.
}
