% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quant.R
\docType{class}
\name{Quant-class}
\alias{Quant-class}
\alias{Quant}
\alias{is.Quant}
\alias{predx_to_json,Quant-method}
\alias{as.data.frame,Quant-method}
\alias{transform_predx,Quant-method}
\title{Quant class: Predictions specified as a set of quantiles}
\usage{
Quant(x)

is.Quant(x)

\S4method{predx_to_json}{Quant}(x)

\S4method{as.data.frame}{Quant}(x, row.names = NULL, optional = FALSE, ...)

\S4method{transform_predx}{Quant}(x, to_class, ...)
}
\description{
A \code{predx} class for binned probabilistic predictions with sequential numeric bins specified by the (inclusive) lower bound of each bin.
}
\details{
Bins are equally spaced and specified by an inclusive lower bound (\code{lwr}). If the bin width is 0.1, \code{lwr} could be 0, 0.1, 0.2, etc. in which case the bins are 0 <= x < 0.1, 0.1 <= x < 0.2, etc. The non-inclusive upper bound of the final bin is assumed to be the maximum value of \code{lwr} plus the uniform bin width. Individual probabilities (\code{prob}) must be greater than or equal to 0 and less than or equal to 1 and the vector of probabilities must sum to 1.

\code{lwr} and \code{prob} must be paired when a BinLwr object is created, but do not need to be in the order of \code{lwr}. The BinLwr object will order them.

In JSON and CSV representations, the bins are named \code{lwr} and the probabilities are named \code{prob}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{A numeric matrix with two columns: \code{lwr} and \code{prob}.}
}}

