% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\docType{class}
\name{Sample-class}
\alias{Sample-class}
\alias{Sample}
\alias{is.Sample}
\alias{as.list,Sample-method}
\alias{as.data.frame,Sample-method}
\alias{transform_predx,Sample-method}
\title{Sample class: Samples from a predictive distribution}
\usage{
Sample(x)

is.Sample(x)

\S4method{as.list}{Sample}(x, ...)

\S4method{as.data.frame}{Sample}(x, row.names = NULL, optional = FALSE,
  ...)

\S4method{transform_predx}{Sample}(x, to_class, cat = NULL, ...)
}
\description{
A \code{predx} class for samples from a predictive distribution.
}
\details{
For now, the samples must be for a quantitative variable.

In JSON and CSV representations, the samples are named \code{sample}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{A vector of samples.}
}}

