% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCat.R
\docType{class}
\name{SampleCat-class}
\alias{SampleCat-class}
\alias{SampleCat}
\alias{is.SampleCat}
\alias{as.list,SampleCat-method}
\alias{as.data.frame,SampleCat-method}
\alias{transform_predx,SampleCat-method}
\title{SampleCat class: Samples from a predictive distribution over categorical bins.}
\usage{
SampleCat(x)

is.SampleCat(x)

\S4method{as.list}{SampleCat}(x, ...)

\S4method{as.data.frame}{SampleCat}(x, row.names = NULL,
  optional = FALSE, ...)

\S4method{transform_predx}{SampleCat}(x, to_class, cat, ...)
}
\description{
A \code{predx} class for samples from a predictive distribution over
categorical bins.
}
\details{
In JSON and CSV representations, the samples are named \code{sample}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{A list with two entries: \code{cat} (character) and \code{sample}
(character).  \code{cat} contains all possible categorical bins and
\code{sample} is a vector of samples of those bins.}
}}

