% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_predx.R
\name{transform_predx}
\alias{transform_predx}
\title{Transform a \code{predx} objects to a target predx class.}
\usage{
transform_predx(x, class, cat, lwr)
}
\arguments{
\item{x}{a \code{predx} object or list of \code{predx} objects}

\item{class}{character specifying \code{predx} class to convert to, e.g. "BinLwr"}

\item{cat}{if class is "BinCat", character vector of categorical bin names}

\item{lwr}{if class is "BinLwr", numeric vector of lower bounds for bins}
}
\value{
A \code{predx} object or list of \code{predx} objects, with NAs and warning messages for failed conversions.
}
\description{
Transform a \code{predx} objects to a target predx class.
}
\details{
This function converts a \code{predx} object or list of \code{predx}
objects to a specified \code{predx} class.  Currently, the only supported
transformations are from Sample to BinLwr and from SampleCat to BinCat.
}
