% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_expected.R
\name{verify_expected}
\alias{verify_expected}
\title{Verify inclusion of expected predictions}
\usage{
verify_expected(x, expected_list, return_df = F)
}
\arguments{
\item{x}{predx_df}

\item{expected_list}{list of lists (see Note)}

\item{return_list}{if \code{TRUE} will return a list of missing predictions}
}
\value{
If any forecasts in \code{expected_list} are not found, the function returns
a list of data.frames of combinations of expected variables for which no
predictions were present in \code{x}.
}
\description{
Function to check a set of predictions for a set of expected predictions
(e.g. targets, locations)
}
\note{
\code{expected_list} is a two-level list of sets of expected predictions.
The lower level (e.g. \code{expected_list[[1]]}) is a list of named character vectors
with specific expected predictions. The names should match column names
in \code{x}. The function checks that all combinations of those elements are
present in \code{x}. A named vector \code{predx_class} may be used to check
predx types.

Different sets in the first level list (e.g.
\code{expected_list[[1]]} and \code{expected_list[[2]]}) are checked separately.
}
\examples{
predx_demo <- predx_df(list(
 location = c('Mercury', 'Venus', 'Earth'),
 target = 'habitability',
 predx = as.Binary(c(1e-4, 1e-4, 1))
))

expected_demo <- list(
  list(
    location = c('Mercury', 'Venus'),
    target = 'habitability',
    predx_class = 'Binary')
  )

expected_demo2 <- list(
  list(
    location = c('Mercury', 'Mars'),
    target = 'habitability',
    predx_class = 'Binary')
  )

verify_expected(predx_demo, expected_demo)
verify_expected(predx_demo, expected_demo2)
}
