% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aedes.R
\name{aedes}
\alias{aedes}
\alias{import_aedes_csv}
\alias{export_aedes_csv}
\alias{aedes_expected}
\title{Tools for working with Aedes challenge forecasts}
\usage{
import_aedes_csv(file, add_vars = NULL)

export_aedes_csv(x, filename = NULL, overwrite = F)

aedes_expected()
}
\arguments{
\item{file}{a csv file formatted for the EPI Aedes Forecasting Challenge}
}
\description{
These functions are used to convert Epidemic Prediction Initiative (EPI) Aedes Forecasting Challenge-formatted csv and JSON files to `predx_df` objects.
}
\examples{
x = import_aedes_csv('vignettes/aedes_null_forecast.csv',
  add_vars=list(due_date='2019-03-31', team='null'))
}
