% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary.R
\docType{class}
\name{Binary-class}
\alias{Binary-class}
\alias{Binary}
\alias{is.Binary}
\alias{predx_to_json,Binary-method}
\alias{as.data.frame,Binary-method}
\alias{transform_predx,Binary-method}
\title{Binary class: Binary predictions}
\usage{
Binary(x)

is.Binary(object)

\S4method{predx_to_json}{Binary}(x)

\S4method{as.data.frame}{Binary}(x, row.names = NULL, optional = FALSE,
  ...)

\S4method{transform_predx}{Binary}(x, to_class, ...)
}
\description{
A \code{predx} class for binary probabilistic predictions.
}
\details{
A single numeric probability that is greater than or equal to 0 and less than or equal to 1.

In JSON and CSV representations, this probability is named \code{prob}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{A single numeric probability.}
}}

