% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary.R
\docType{class}
\name{Binary-class}
\alias{Binary-class}
\title{Binary predictions}
\description{
This predx class is used to capture binary probabilistic predictions. It contains a single numeric probability that is greater than or equal to 0 and less than or equal to 1. In JSON and CSV representations, this probability is named 'prob'.
}
\section{Slots}{

\describe{
\item{\code{predx}}{Contains a single numeric probability (0 <= p <= 1).}
}}

