% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinCat.R
\docType{class}
\name{BinCat-class}
\alias{BinCat-class}
\alias{BinCat}
\alias{is.BinCat}
\alias{predx_to_json,BinCat-method}
\alias{as.data.frame,BinCat-method}
\alias{transform_predx,BinCat-method}
\title{BinCat class: Binned predictions with categorical bins}
\usage{
BinCat(x)

is.BinCat(x)

\S4method{predx_to_json}{BinCat}(x)

\S4method{as.data.frame}{BinCat}(x, row.names = NULL, optional = FALSE,
  ...)

\S4method{transform_predx}{BinCat}(x, to_class, ...)
}
\description{
A \code{predx} class for binned probabilistic predictions with categorical bins specified by strings.
}
\details{
\code{BinCat} is flexible as it can be used for any binned predictions. Individual probabilities (\code{prob}) must be greater than or equal to 0 and less than or equal to 1 and the vector of probabilities must sum to 1.

In JSON and CSV representations, the bins are named \code{cat} and the probabilities are named \code{prob}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{A data.frame with two columns: \code{cat} (character) and \code{prob} (numeric).}
}}

