% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_csv.R
\name{export_csv}
\alias{export_csv}
\title{Title}
\usage{
export_csv(x, filename = NULL, overwrite = F)
}
\arguments{
\item{filename}{Path or file. If NA, data.frame is returned.}

\item{overwrite}{Overwrite the file if it already exists? Default is FALSE.}

\item{forecast}{predx_df object.}
}
\description{
Title
}
