% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_json}
\alias{export_json}
\title{Title}
\usage{
export_json(x, filename = NA, overwrite = F)
}
\arguments{
\item{filename}{character. Filename to write JSON file. If NA, JSON is returned.}

\item{overwrite}{boolean. Overwrite the file if it already exists.}

\item{forecast}{predx_df.}
}
\description{
Title
}
\examples{
export_json(x, 'inst/data/flusight.json')
}
