% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.R
\docType{class}
\name{Point-class}
\alias{Point-class}
\alias{Point}
\alias{is.Point}
\alias{as.list,Point-method}
\alias{as.data.frame,Point-method}
\title{Point class: Point predictions}
\usage{
Point(x)

is.Point(object)

\S4method{as.list}{Point}(x, ...)

\S4method{as.data.frame}{Point}(x, row.names = NULL, optional = FALSE,
  ...)
}
\description{
This predx class is used to capture point predictions. It contains a single numeric point predictions with no other restrictions. In JSON and CSV representations, this value is named \code{point}.
}
\section{Slots}{

\describe{
\item{\code{predx}}{Contains a single numeric point prediction.}
}}

\examples{

}
