% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.R
\docType{class}
\name{Point-class}
\alias{Point-class}
\title{Point predictions}
\description{
This predx class is used to capture point predictions. It contains a single numeric point predictions with no other restrictions. In JSON and CSV representations, this value is named 'point'.
}
\section{Slots}{

\describe{
\item{\code{predx}}{Contains a single numeric point prediction.}
}}

\examples{

}
