% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_df.R
\name{predx_df}
\alias{predx_df}
\alias{validate_predx_df}
\title{Title}
\usage{
validate_predx_df(x, return = F)
}
\description{
Title
}
\examples{
as.predx_df(list(
 location = c('Mercury', 'Venus', 'Earth'),
 target = 'habitability',
 predx = list(Binary(0), Binary(0), Binary(1))
))
}
