% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predx_df.R
\name{predx_df}
\alias{predx_df}
\alias{validate_predx_df}
\alias{is.predx_df}
\alias{as.predx_df}
\title{Tools to check the \code{predx} data frames contain valid \code{predx_class} and \code{predx} columns}
\usage{
validate_predx_df(x, return = F)

is.predx_df(x)

as.predx_df(x)
}
\value{
Error message or TRUE.
}
\description{
Tools to check the \code{predx} data frames contain valid \code{predx_class} and \code{predx} columns
}
\examples{
as.predx_df(list(
 location = c('Mercury', 'Venus', 'Earth'),
 target = 'habitability',
 predx = list(Binary(0), Binary(0), Binary(1))
))
}
