% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCat.R
\docType{class}
\name{SampleCat-class}
\alias{SampleCat-class}
\alias{SampleCat}
\alias{is.SampleCat}
\alias{predx_to_json,SampleCat-method}
\alias{as.data.frame,SampleCat-method}
\alias{transform_predx,SampleCat-method}
\title{SampleCat class: Samples from a distribution of categories or classifications.}
\usage{
SampleCat(x)

is.SampleCat(x)

\S4method{predx_to_json}{SampleCat}(x)

\S4method{as.data.frame}{SampleCat}(x, row.names = NULL,
  optional = FALSE, ...)

\S4method{transform_predx}{SampleCat}(x, to_class, cat, ...)
}
\description{
A \code{predx} class for samples from a distribution of
categories or classifications specified as character strings.
}
\details{
In JSON and CSV representations, the samples are named \code{sample}.
}
\examples{
x <- SampleCat(sample(letters, size = 10))
x

}
