% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_predx.R
\name{transform_predx}
\alias{transform_predx}
\alias{transform_predx,list-method}
\title{Transform a \code{predx} objects to a target predx class.}
\usage{
transform_predx(x, to_class, ...)

\S4method{transform_predx}{list}(x, to_class, ...)
}
\arguments{
\item{x}{a \code{predx} object or list of \code{predx} objects}

\item{to_class}{character specifying \code{predx} class to convert to, e.g. "BinLwr"}

\item{cat}{if to_class is "BinCat", character vector of categorical bin names}

\item{lwr}{if to_class is "BinLwr", numeric vector of lower bounds for bins}
}
\value{
A \code{predx} object or list of \code{predx} objects, with NAs and warning messages for failed conversions.
}
\description{
Transform a \code{predx} objects to a target predx class.
}
\details{
This function converts a \code{predx} object or list of \code{predx}
objects to another specified \code{predx} class when possible.
}
